fix test flakiness by using `afterRender` to schedule active state
Before, both the `flash-message` component and its corresponding unit
test used `Ember.run.next`. under the hood, `Ember.run.next`
essentially uses `setTimeout(workFn, 1)`. One issue you can have with
this (and I think we saw with flaky tests on CI), is that browsers are
not required and usually do not have precise timing or ordering when
running `setTimeout` hooks. What could have been happening with the
flaky test errors is that the `Ember.run.next` in the test finished
_before_ the `Ember.run.next` that was setup in the `flash-message`
component.

Now, we use `Ember.run.scheduleOnce('afterRender')` instead.
`afterRender` adds the function to the _existing_ run loop vs
`Ember.run.next`, which always runs in a new run loop. Because the work
is run in the same run loop (but after rendering is complete) and
doesn't have any asynchronous side effects, we no longer need the
`Ember.run.next` in the tests.